Build : 빌드 에러 해결을 위한 컴포넌트 구조 분리
